Dynamic delete now working (persists to db, removes rows dynamically)

+++ sqk/datasets/views.py
@@ -3,7 +3,8 @@ from django.views.generic import View, ListView, DetailView, FormView, UpdateVie
 from django.views.generic.detail import SingleObjectMixin
 from django.core.urlresolvers import reverse, reverse_lazy
 from django.conf import settings
-from django.shortcuts import get_object_or_404
+from django.shortcuts import get_object_or_404, render_to_response
+from django.template import RequestContext
 from django.utils import simplejson
 from django.core.serializers import serialize
 from django.http import HttpResponse, HttpResponseRedirect
@@ -12,7 +13,7 @@ from sqk.datasets.forms import DatasetForm, DatasetEditForm, DatasourceForm, Lab
 from sqk.datasets.models import *
 from sqk.datasets.tasks import read_datasource, handle_uploaded_file, extract_features
 
-## Dataset views
+############## Dataset views ##################
 class DatasetList(ListView):
     model = Dataset
     queryset = Dataset.objects.order_by('-created_at')
@@ -112,7 +113,8 @@ def delete_dataset(request, pk):
     dataset.delete()
     return HttpResponseRedirect(reverse('datasets:index'))
 
-## Instance views
+################ Instance views ####################
+
 class InstanceDetail(DetailView):
     model = Instance
     context_object_name = 'instance'
@@ -173,16 +175,24 @@ def delete_instances(request, dataset_id):
 
     Must be a POST request.
     '''
-    results = {'success': False}
     # Get the ids of the selected instances
-    instance_ids = [int(instance_id) for instance_id in request.POST.getlist('selected[]')]
-    for id in instance_ids:
-        inst_obj = Instance.objects.get(pk=id)
-        inst_obj.delete()
-    results['success'] = True
-    instances = Instance.objects.filter(dataset__id=dataset_id)
-    json = serialize('json', instances)
-    return HttpResponse(json, mimetype='application/json')
+    if request.is_ajax():
+        instance_ids = [int(instance_id) for instance_id in request.POST.getlist('selected[]')]
+        for id in instance_ids:
+            inst_obj = Instance.objects.get(pk=id)
+            inst_obj.delete()
+        return HttpResponse(simplejson.dumps("success"), mimetype='application/json')
+    else:
+        # TODO: handle non-AJAX
+        return HttpResponseRedirect(reverse('datasets:detail', args=(dataset_id,)))
+
+def ajax_test(request):
+    if request.is_ajax():
+        message = "This is an Ajax message"
+    else:
+        message = "Ajax error"
+    return render_to_response("datasets/ajax.html",
+        context_instance=RequestContext(request,{'message':message}))
 
 def update_instances_labels(request, dataset_id, label_name_id):
     '''View for updating the label values for selected instances.
@@ -234,7 +244,7 @@ class LabelNameCreate(FormView):
         return reverse_lazy('datasets:detail',
             kwargs={'pk': self.kwargs['dataset_id']})
 
-# X-editable views
+###################### X-editable views ##################
 
 @ensure_csrf_cookie
 def update_name(request, dataset_id):
